fastapi_site/requirements.txt
fastapi
uvicorn[standard]
jinja2
pyyaml
plotly
markdown
reportlab
